Some small improvements.
