+++ source/opengl.h
@@ -29,6 +29,7 @@ class OpenGL
 {
 public:
 	static bool HasAdaptiveVSyncSupport();
+	static bool HasSwizzleSupport();
 };
 
 #endif
